Time: 4 ms (7.31%) | Memory: 19.7 MB (35.62%) - LeetSync

+++ 662-maximum-width-of-binary-tree/maximum-width-of-binary-tree.cpp
@@ -12,8 +12,8 @@
 class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
-       queue<pair<TreeNode*,int>> q;
-       q.push({root,0});
+       queue<pair<TreeNode*,long long>> q;
+       q.push({root,0LL});
        int width=0;
        while(!q.empty()){
         int size=q.size();
